% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCLT.R
\name{makeCLT}
\alias{makeCLT}
\alias{is.CLT}
\alias{lxFun}
\alias{lxFun.default}
\alias{lxFun.CLT}
\alias{axFun}
\alias{axFun.default}
\alias{axFun.CLT}
\alias{dxFun}
\alias{dxFun.default}
\alias{dxFun.CLT}
\alias{qxFun}
\alias{qxFun.default}
\alias{qxFun.CLT}
\title{Cohort Life Table Constructor}
\usage{
makeCLT(n, m)

\method{is}{CLT}(x)

lxFun(x, ...)

\method{lxFun}{default}(n_x)

\method{lxFun}{CLT}(x)

axFun(x, ...)

\method{axFun}{default}(n_vec)

\method{axFun}{CLT}(x)

dxFun(x, ...)

\method{dxFun}{default}(n_vec)

\method{dxFun}{CLT}(x)

qxFun(x, ...)

\method{qxFun}{default}(a_vec)

\method{qxFun}{CLT}(x)
}
\arguments{
\item{n}{A vector of cohort-specific sizes/number of individuals. Should
include a last cohort with n=0, else the function adds this.}

\item{m}{A vector of cohort-specific fecundities. Should be of length
\code{length(n)-1} or else have \code{NA} at \code{n[length(n)]=0}.}

\item{x}{An object of class \sQuote{CLT}.}

\item{n_x}{A vector containing the number of individuals in each cohort}

\item{a_x}{A vector containing Annual survival for each cohort}
}
\value{
A \code{vector} or \code{data.frame} of class \sQuote{CLT} containing
  one or all of \code{names}:
  \describe{
    \item{x }{Cohort ages/classes}
    \item{n_x }{Number of individuals in each cohort}
    \item{l_x }{Survivorship}
    \item{a_x }{Annual survival}
    \item{d_x }{Number of individuals that die}
    \item{q_x }{Annual mortality}
    \item{m_x }{Fecundity}
  }
}
\description{
Given age-specific cohort size (n) and fecundity (m), construct a cohort life
table of class \sQuote{CLT}.
}
\examples{
  makeCLT(n = c(10, 5, 3, 2, 0), m = c(0, 1, 3, 2))
}
\author{
\email{matthewwolak@gmail.com}
}
