% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCaccessories.R
\name{postPlot}
\alias{postPlot}
\title{Posterior distribution plot.}
\usage{
postPlot(posterior, bw = "nrd", plotHist = TRUE, histbreaks = 100,
  prior = NULL, prange = c("prior", "posterior"), main, sub, ylim,
  denslwd = 6, denslty = "solid", denscol = "black", hpdlwd = 6,
  hpdlty = "dashed", hpdcol = "black", meanlwd = 7, meanlty = "12",
  meancol = "red", priorlwd = 3, priorlty = "solid", priorcol = "green",
  at1 = NULL, at2 = NULL, labels1 = NULL, labels2 = NULL, plot = TRUE,
  ...)
}
\arguments{
\item{posterior}{Object containing a marginal posterior distribution from
from which kernel density estimates can be calulated by the
\code{coda::density} function.}

\item{bw}{Either a \code{numeric} or \code{character} for calculating the
smoothing bandwidth used in the kernel density estimation. If
\code{numeric}, the value is used as the standard deviation of the
smoothing kernel. If \code{character}, the string is used as a standard
rule to choose the bandwith. See details in \code{\link[stats]{density}}.
Note that the default value \code{"nrd"} is not the default used in
\code{\link[stats]{density}}.}

\item{plotHist}{Should a histogram of the density be plotted as well.}

\item{histbreaks}{If \code{plotHist = TRUE}, then the number of breaks. See
\code{\link[graphics]{hist}} for details.}

\item{prior}{The prior distribution for a line be added to the plot. If
\code{NULL} (the default) then no line is added. A prior is plotted if
either a \code{\link[coda]{mcmc}} object or \code{list} with the prior 
density is supplied. See Details.}

\item{prange}{Character argument specifying the relative range
\code{c("prior", "posterior")} for the density estimation of the prior.
The supplied argument must begin with either \code{prior} or
\code{posterior} exactly, but can end with a numeric value indicating the
relative range of the prior posterior to be used. See Details. Range
adjustment currently only works when class \code{mcmc} provided to
\code{prior}.#FIXME}

\item{main}{Overall title for the plot. See \code{\link[graphics]{plot}} and
\code{\link[graphics]{title}}.}

\item{sub}{Sub-title for the plot. See \code{\link[graphics]{plot}} and
\code{\link[graphics]{title}}.}

\item{ylim}{Y-axis limits to the plotting region, passed to
\code{\link[coda]{densplot}}. See \code{\link[graphics]{plot.window}}.}

\item{denslwd, hpdlwd, meanlwd, priorlwd}{The line widths for the lines to be plotted
representing the kernel density estimate (\code{denslwd}), credible
interval limits (\code{hpdlwd}), or posterior mean (\code{meanlwd}). See
\code{\link[graphics]{par}} for details.}

\item{denslty, hpdlty, meanlty, priorlty}{The line type for the lines to be plotted
representing the kernel density estimate (\code{denslty}), credible
interval limits (\code{hpdlty}), or posterior mean (\code{meanlty}). See
\code{\link[graphics]{par}} for details.}

\item{denscol, hpdcol, meancol, priorcol}{The line colors for the lines to be plotted
representing the kernel density estimate (\code{denscol}), credible
interval limits (\code{hpdcol}), or posterior mean (\code{meancol}). See
\code{\link[graphics]{par}} for details.}

\item{at1, at2}{The points at which tick-marks are to be drawn on the
x-axis (\code{at1}) and y-axis (\code{at2}). See
\code{\link[graphics]{axis}} for details.}

\item{labels1, labels2}{Labels to be placed at the tick-marks of the
x-axis (\code{labels1}) and and y-axis (\code{labels2}). See
\code{\link[graphics]{axis}} for details.}

\item{plot}{Logical. If \code{TRUE} (default) the plotting is performed,
otherwise the graphical output is suppressed.}

\item{\dots}{Additional arguments passed to \code{densplot}, \code{hist},
\emph{and} \code{abline}.}
}
\value{
A \code{list} returned invisibly (assign to an object to view):
  \describe{
    \item{call}{The original \code{call}.}
    \item{postDensity}{List containing objects describing the posterior
 distribution. Items include a \code{character} indicating if the
 density was adjusted because of boundary constraints, an object of
 \code{class} "density", and an object of \code{class} "histogram".}
    \item{priorDensity}{List containing objects describing the prior
 distribution. Items include a \code{character} indicating if the
 density was adjusted because of boundary constraints and an object of
	 \code{class} "density".}
  }
}
\description{
Pretty plot of a posterior distribution that can include summary statistics,
such as the posterior mean and credible interval as well as a prior density.
}
\details{
To plot a prior distribution, \code{prior} should either specify an object of
  class \code{\link[coda]{mcmc}} that contains a simulated prior distribution
  or it should contain a \code{list} which contains the prior density (where
  the first two elements must be named \code{x} and \code{y} and contain the
  quantiles and densities, respectively.). Note that a function to generate the
  prior distribution may be specified. However, this must return either a
  \code{mcmc} or \code{list} object.

TODO: Add more about prior plotting, specifically how parameter expanded
  priors are simulated versus inverse-Wishart/Gamma are derived straight from
  their distribution functions. Also, note how density is calculated#TODO

If the prior is of class \code{mcmc} then the argument \code{prange} can be
  used to either extend or restrict the range over which the prior density
  estimation is conducted, in relation to the posterior distribution.
  The range alteration is done by specifying \code{posteriorN}, or appending
  the \code{"posterior"} character string with a numeric value indicating
  the desired adjustment of the range. Prior density will then be calculated
  after taking a subset of the current samples that fall in the range. Note,
  this process only makes sense when used in conjunction with
  \code{posterior}. By default, a \code{1} is used if there is no numeric
  suffix; indicating that a 1-to-1 mapping of the range should be used.
  Values of the suffix >1 will extend the range, while those <1 will contract
  it.
}
\section{Warning}{

The use of the \dots argument is untested. Arguments of the same name that
are shared among \code{densplot}, \code{hist}, or \code{abline} will affect
each of these aspects of the plot if supplied to \code{postPlot}. Proceed
with caution!
}

\examples{
# Simulate two example MCMC chains:
## Both are standard normal distributions of only positive values
## However, one is on the boundary (near zero) while the other is not
normMCMC <- coda::mcmc(matrix(c(abs(rnorm(1000, 0.2, 1)), rnorm(1000, 10, 1)),
ncol = 2))
par(mfrow = c(2, 1))
  postPlot(normMCMC[, 1], ylim = c(0, 1))
  postPlot(normMCMC[, 2], ylim = c(0, 1))

# Now include an inverse Gamma prior, using the density function
postPlot(normMCMC[, 1], ylim = c(0,1),
  prior = dIW(seq(1e-16, 5, length = 1000), V = diag(1), nu = 1))

# Example with parameter expanded prior:
pepr <- coda::mcmc(rpeIW(1000, V = diag(2), nu = 2, alpha.mu = rep(0, 2),
  alpha.V = diag(2)*1000))
par(mfrow = c(2, 1))
  postPlot(normMCMC[, 1], ylim = c(0, 1),
    prior = pepr[,1], prange = "posterior6")
  postPlot(normMCMC[, 2], ylim = c(0, 1),
    prior = pepr[,4], prange = "posterior1.5") 
# different scalar to get same y-intercept

}
\seealso{
\code{\link[coda]{densplot}}, \code{\link[graphics]{hist}}

Other MCMC posterior distribution helper functions: \code{\link{plot2mcmc}},
  \code{\link{postTable}}
}
\author{
\email{matthewwolak@gmail.com}
}
