% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCaccessories.R
\name{postTable}
\alias{postTable}
\title{Posterior distribution summary.}
\usage{
postTable(mcpost, ind = NULL, sigdig = 3, ...)
}
\arguments{
\item{mcpost}{Object of class \code{mcmc} containing marginal posterior
distributions.}

\item{ind}{Column indices for the subset of \code{mcpost} for which the
summary statistics are to be returned.}

\item{sigdig}{Integer value for the number of significant digits to report.}

\item{\dots}{Additional arguments.}
}
\value{
A character \code{matrix} of two rows and \code{ncol(mcpost)} columns
  (\code{length(ind)} when \code{ind} is not \code{NULL}). The first row
  contains component names from \code{mcpost}. The second row contains a
  a string representing the posterior mode then posterior mean, separated by
  a comma, followed by the lower and upper limits of the highest posterior
  density, enclosed in parentheses.
}
\description{
Reports a table containing the mode, mean, and highest posterior density of
the marginal posterior distribution.
}
\examples{
# Simulate to example MCMC chains:
## Both are standard normal distributions of only positive values
## However, one is on the boundary (near zero) while the other is not
normMCMC <- coda::mcmc(matrix(c(abs(rnorm(1000, 0.2, 1)), rnorm(1000, 10, 1)),
ncol = 2))
postTable(normMCMC)
}
\seealso{
Other MCMC posterior distribution helper functions: \code{\link{postPlot}}
}
\author{
\email{matthewwolak@gmail.com}
}
