% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_replicate.R
\name{preplicate}
\alias{preplicate}
\title{Parallel Implementation of replicate.}
\usage{
preplicate(n, expr, simplify = "array", ...)
}
\arguments{
\item{n}{integer: the number of replications.}

\item{expr}{the expression (a language object, usually a call) to evaluate
repeatedly.}

\item{simplify}{logical or character string, see \code{replicate}.}

\item{...}{other arguments to be passed to \code{mclapply}.}
}
\value{
A list, vector, or matrix depending on arguments to \code{simplify}.
  See documentation for \code{replicate} for more details.
}
\description{
Implements parallel version of \code{replicate} using
\code{parallel::mclapply()}.
}
\examples{
\dontrun{
  system.time(r <- replicate(10000, var(rnorm(500))))
  system.time(pr <- preplicate(10000, var(rnorm(500)), mc.cores = 3))
}
}
\author{
\email{matthewwolak@gmail.com}
}
