% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{rpeIW}
\alias{rpeIW}
\title{Parameter expanded inverse Wishart (co)variance.}
\usage{
rpeIW(n = 1, V, nu, alpha.mu, alpha.V)
}
\arguments{
\item{n}{Integer of how many samples to generate.}

\item{V}{Numeric matrix of the expected (co)variances}

\item{nu}{Numeric for the degree of belief parameter for the inverse-
Wishart.}

\item{alpha.mu}{Numeric vector of means for the redundant working parameters.}

\item{alpha.V}{Numeric matrix of the covariance matrix for the redunant
working parameters.}
}
\value{
A numeric vector (if the dimensions of \code{V} are 1) or matrix
  where each row designates a sampled (co)variance and the columns contain
  the matrix elements for the prior covariance matrix.
}
\description{
Simulate prior (co)variance matrix according to the inverse Wishart
distribution with parameter expansion on the hyperparameters.
}
\details{
Details on the meaning of \code{V, nu, alpha.mu, alpha.V} can be found
in the \code{\link[MCMCglmm]{MCMCglmm}} documentation covering the way
to specify variance structure priors in the \code{prior} argument.
}
\examples{
set.seed(101)
(peIW <- rpeIW(n = 5, V = diag(2), nu = 3,
alpha.mu = rep(0, 2), alpha.V = diag(2)*1000))
sapply(1:5, FUN = function(i){matrix(peIW[i, ], 2, 2)}, simplify = "array")
}
\seealso{
\code{\link[MCMCglmm]{MCMCglmm}}, \code{\link[MCMCglmm]{rIW}}

Other prior functions: \code{\link{dIW}}
}
\author{
\email{matthewwolak@gmail.com}
}
