% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{wolak_palette}
\alias{wolak_palette}
\title{Create wolakR package palette}
\usage{
wolak_palette(name = "aucfriend", n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Character specifying name of palette to use. See
\code{\link{wolak_palettes}} for list of choices.}

\item{n}{Integer number of colors to return.}

\item{type}{Character indicating either \dQuote{discrete} for colors already
listed in the paletter or \dQuote{continuous} to interpolate between colors.}
}
\value{
A vector of colors of the class \dQuote{palette} with a palette name.
}
\description{
Sets up the R object of class palette that contains one of the pre-specified
  color palettes.
}
\details{
Function is based off the \code{wesanderson::wes_palette} function.
}
\examples{
clr <- wolak_palette("aucfriend")
\dontrun{
  # check:
    colorblindcheck::palette_check(clr, plot = TRUE)
}
}
\author{
\email{matthewwolak@gmail.com}
}
