% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wolak_palettes}
\alias{wolak_palettes}
\title{List of all color palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
wolak_palettes
}
\description{
Run the function \code{\link{wolak_palette}} to create the palette objects.
}
\keyword{datasets}
