% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCaccessories.R
\name{postPlot}
\alias{postPlot}
\title{Posterior distribution plot.}
\usage{
postPlot(posterior, plotHist = TRUE, histbreaks = 100, prior = NULL,
  prange = c("prior", "posterior"), main, sub, ylim, denslwd = 6,
  denslty = "solid", denscol = "black", hpdlwd = 6, hpdlty = "dashed",
  hpdcol = "black", meanlwd = 7, meanlty = "12", meancol = "red",
  priorlwd = 3, priorlty = "solid", priorcol = "green", at1 = NULL,
  at2 = NULL, labels1 = NULL, labels2 = NULL, ...)
}
\arguments{
\item{posterior}{Object containing a marginal posterior distribution from
from which kernel density estimates can be calulated by the
\code{coda::density} function.}

\item{plotHist}{Should a histogram of the density be plotted as well.}

\item{histbreaks}{If \code{plotHist = TRUE}, then the number of breaks. See
\code{\link[graphics]{hist}} for details.}

\item{prior}{The prior distribution for a line be added to the plot. If
\code{NULL} (the default) then no line is added. A prior is plotted if
either a \code{\link[coda]{mcmc}} object or \code{list} with the prior 
density is supplied. See Details.}

\item{prange}{Character argument specifying which range
\code{c("prior", "posterior")} the prior distribution should span. Currently,
only works when class \code{mcmc} provided to \code{prior}.#FIXME}

\item{main}{Overall title for the plot. See \code{\link[graphics]{plot}} and
\code{\link[graphics]{title}}.}

\item{sub}{Sub-title for the plot. See \code{\link[graphics]{plot}} and
\code{\link[graphics]{title}}.}

\item{ylim}{Y-axis limits to the plotting region, passed to
\code{\link[coda]{densplot}}. See \code{\link[graphics]{plot.window}}.}

\item{denslwd, hpdlwd, meanlwd, priorlwd}{The line widths for the lines to be plotted
representing the kernel density estimate (\code{denslwd}), credible
interval limits (\code{hpdlwd}), or posterior mean (\code{meanlwd}). See
\code{\link[graphics]{par}} for details.}

\item{denslty, hpdlty, meanlty, priorlty}{The line type for the lines to be plotted
representing the kernel density estimate (\code{denslty}), credible
interval limits (\code{hpdlty}), or posterior mean (\code{meanlty}). See
\code{\link[graphics]{par}} for details.}

\item{denscol, hpdcol, meancol, priorcol}{The line colors for the lines to be plotted
representing the kernel density estimate (\code{denscol}), credible
interval limits (\code{hpdcol}), or posterior mean (\code{meancol}). See
\code{\link[graphics]{par}} for details.}

\item{at1, at2}{The points at which tick-marks are to be drawn on the
x-axis (\code{at1}) and y-axis (\code{at2}). See
\code{\link[graphics]{axis}} for details.}

\item{labels1, labels2}{Labels to be placed at the tick-marks of the
x-axis (\code{labels1}) and and y-axis (\code{labels2}). See
\code{\link[graphics]{axis}} for details.}

\item{\dots}{Additional arguments passed to \code{densplot}, \code{hist},
\emph{and} \code{abline}.}
}
\value{
A \code{list} returned invisibly (assign to keep):
  \describe{
    \item{call}{The original \code{call}.}
    \item{postDensity}{Object of \code{class} "density" for the posterior
      distribution.}
    \item{priorDensity}{Object of \code{class} "density" for the prior
      distribution.}
    \item{bandwith}{The density bandwith from the posterior used to set some
        the bandwith in the prior under some constraints.FIXME}
    \item{histOut}{Object of \code{class} "histogram" for the posterior
        distribution.}
    \item{constraint}{A \code{character} indicating if any densities were
        adjusted because of boundary constraints.}
  }
}
\description{
Pretty plot of a posterior distribution that can include summary statistics,
such as the posterior mean and credible interval as well as a prior density.
}
\details{
To plot a prior distribution, \code{prior} should either specify an object of
  class \code{\link[coda]{mcmc}} that contains a simulated prior distribution
  or it should contain a \code{list} which contains the prior density (where
  the first two elements must be named \code{x} and \code{y} and contain the
  quantiles and density, respectively.). Note that a function to generate the
  prior distribution may be specified. However, this must return either a
  \code{mcmc} or \code{list} object.

Add more about prior plotting, specifically how parameter expanded priors are
  simulated versus inverse-Wishart/Gamma are derived straight from their
  distribution functions. Also, note how density is calculated
  after 'trimming' to range of the posterior samples.
}
\section{Warning}{

The use of the \dots argument is untested. Arguments of the same name that
are shared among \code{densplot}, \code{hist}, or \code{abline} will affect
each of these aspects of the plot if supplied to \code{postPlot}. Proceed
with caution!
}

\examples{
# Simulate to example MCMC chains:
## Both are standard normal distributions of only positive values
## However, one is on the boundary (near zero) while the other is not
normMCMC <- coda::mcmc(matrix(c(abs(rnorm(1000, 0.2, 1)), rnorm(1000, 10, 1)),
ncol = 2))
par(mfrow = c(2, 1))
  postPlot(normMCMC[, 1], ylim = c(0, 1))
  postPlot(normMCMC[, 2], ylim = c(0, 1))

# Now include an inverse Gamma prior, using the density function
postPlot(normMCMC[, 1], ylim = c(0,1),
  prior = dIW(seq(1e-16, 5, length = 1000), V = diag(1), nu = 1))
}
\seealso{
\code{\link[coda]{densplot}}, \code{\link[graphics]{hist}}

Other MCMC posterior distribution helper functions: \code{\link{postTable}}
}
\author{
\email{matthewwolak@gmail.com}
}
